IChangement des paramètres de JpaRepository de Jour et Disponibilite

+++ src/main/java/com/univtime/informatique/repositories/DisponibiliteRepository.java
@@ -3,6 +3,6 @@ package com.univtime.informatique.repositories;
 import com.univtime.informatique.entities.DisponibiliteEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DisponibiliteRepository extends JpaRepository<DisponibiliteEntity, Long> {
+public interface DisponibiliteRepository extends JpaRepository<DisponibiliteEntity, Integer> {
 
 }
